pyproject.toml
[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = [{ index = "pytorch-gpu" }]

[project]
name = "deepfake-detection-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "mediapipe>=0.10.31",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.2.1",
]
